# src/AggStudentDiscounts.Api/openapi.yaml
openapi: 3.0.3
info:
  title: API Скидки студентов (карта заведений)
  description: |
    REST API для проекта интерактивной карты скидок для студентов.
    Поддерживает роли: Гость, Пользователь, Модератор, Администратор.
    Аутентификация через JWT (Bearer token).
  version: 1.0.0
servers:
  - url: https://api.example.com
    description: Продакшн сервер
  - url: http://localhost:5075
    description: Локальный сервер разработки
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Place:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        category:
          type: string
          description: Категория заведения (кафе, магазин и т.д.)
        discount:
          type: string
          description: Описание скидки
        conditions:
          type: string
          description: Условия получения скидки
        validityPeriod:
          type: string
          nullable: true
          description: Период действия скидки (текст или даты)
        photos:
          type: array
          items:
            type: string
            format: uri
          description: Ссылки на фото (подтверждения)
        sourceUrl:
          type: string
          nullable: true
          description: Ссылка на источник информации
        votesYes:
          type: integer
          description: Количество голосов "Да"
        votesNo:
          type: integer
          description: Количество голосов "Нет"
        lastVoteDateYes:
          type: string
          format: date-time
          nullable: true
          description: Дата последнего голосования "Да"
        lastVoteDateNo:
          type: string
          format: date-time
          nullable: true
          description: Дата последнего голосования "Нет"
        status:
          type: string
          enum: [pending, published, rejected]
          description: Статус публикации (для модерации)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PlaceListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Place'
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer

    # ---------- Заявка на добавление ----------
...
    VoteRequest:
      type: object
      required:
        - vote
      properties:
        vote:
          type: string
          enum: [yes, no]
    VoteStats:
      type: object
      properties:
        yes:
          type: integer
        no:
          type: integer
        lastVoteDateYes:
          type: string
          format: date-time
          nullable: true
        lastVoteDateNo:
          type: string
          format: date-time
          nullable: true
        canVoteAgain:
          type: boolean
          description: Может ли текущий пользователь голосовать (не прошло 15 дней)
        nextVoteDate:
          type: string
          format: date-time
          nullable: true

    # ---------- Фильтры и поиск ----------
...
security:
  - bearerAuth: []
paths:
...
  /api/establishments:
    get:
      summary: Получить список заведений с фильтрацией
      operationId: getEstablishments
      tags:
        - Places
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: district
          schema:
            type: string
        - in: query
          name: radius
          schema:
            type: number
        - in: query
          name: lat
          schema:
            type: number
        - in: query
          name: lng
          schema:
            type: number
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: pageSize
          schema:
            type: integer
      responses:
        '200':
          description: Список заведений (только опубликованные)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceListResponse'
  /api/establishments/{id}:
    get:
      summary: Получить карточку заведения по ID
      operationId: getPlaceById
      tags:
        - Places
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Данные заведения
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Place'
        '404':
          description: Заведение не найдено

  # ----- Голосование за актуальность -----
  /api/establishments/{id}/vote:
    post:
      summary: Проголосовать за актуальность скидки
      operationId: vote
      tags:
        - Voting
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoteRequest'
      responses:
        '200':
          description: Голос учтён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoteStats'
        '429':
          description: Повторное голосование ранее чем через 15 дней
        '401':
          description: Требуется авторизация
  /api/establishments/{id}/votes:
    get:
      summary: Получить статистику голосования по заведению
      operationId: getVoteStats
      tags:
        - Voting
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Статистика голосов и даты последних голосов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoteStats'

  # ----- Заявки от пользователей -----
...